Override elisp-mode-syntax-propertize to leave ## alone
In Emacs 29.1 a syntax property rule was added to treat `##' as the
symbol it is (except in comments and docstrings).  However `##' is
different from other symbols in that no space has to used in between
it and the following symbol, to prevent them from becoming one symbol,
we take advantage of that idiosyncrasy.

We revert this because the feature that depend on `##' having symbol
syntax are less important than the features that would get confused
and treat `##' and the following symbol as one symbol.

However, in docstrings and comments `##' (with the `...' around it)
should be fontified like other symbols, so there we do give it symbol
syntax - unlike the commented out code, which explicitly prevented
that.