Optionally fontify ## and its special arguments
Note that if `llama-fontify-mode' is disabled, then
`font-lock-type-face' is used for optional arguments because the
respective entry in `lisp-el-font-lock-keywords-2' (IMO) matches a bit
too eagerly.  By default `llama-optional-arguments' inherits from that
face and doesn't make any changes, so toggling `llama-fontify-mode',
usually has no effect on the appearance of these arguments.