Make unused arguments at mandatoriness border optional by default
Previously we turned
  (##list %1 &3)
into
  (lambda (%1 _%2 &optional     &3) (list %1 &3))
now we use
  (lambda (%1     &optional _&2 &3) (list %1 &3))

The new approach is better because even if the first form would have
been "correct", the latter will nearly¹ always at least be "valid".

The same isn't true for the old approach.  If we make an argument
mandatory that should be optional, and the callers does not pass that
argument, that's an error.  These cases are several magnitudes more
likely.

¹ The exception being callers that check a functions *minimal* arity
  to determine what arguments to pass.  In the unlikely case that one
  should ever encounter such a caller, one can use an explicit _%N,
  as one previously would frequently have had to use _&N.