llama--collect: Signal error on all argument conflicts
The assert was needlessly limited to the first argument, to catch
specifically `%1'vs`%' and `&1'vs`&' (but also, e.g., `%1'vs`_&').  Now
be also detect `%N'vs`&N', `%N'vs`_%N' and even conflicts were multiple
aspects differ (e.g., `%2'vs`_&2').

We could just use `%N' in the arglist when both `%N' and `_%N' appear
in the body.  Since `_%N' is dropped from the body, that wouldn't result
in an unbound variable error, as would happen in other cases.  But that
would both complicate the implementation and hide a user mistake, which
could be of consequence (e.g., if they actually meant `_%N+1').