llama-font-lock-keywords: Also fontify llama
Use the same face as for `lambda' not the same as for `##'.  The reason
that `##' uses another face is that it might appear right next to
another symbol using `font-lock-keyword-face' with no space in between.
By using another face, we prevent them from looking like one symbol.
The same cannot happen for `llama'.